Add option to use tag wth noenv
The ```--noenv``` flag normally does not check anything out (except for
the build repo).  This makes commands like ```changelog``` useless, as
it acts on the checked out repos.  Specifically, ```changelog```
currently only shows you changes in ```build``` when used with ```-n```

This change adds a flag to immediately check out a tag when using
```--noenv```.