Fix zendev sync stashing
Zendev currently stashes if any of staged/unstaged/untracked are
true.  By default, stash will only store files that are already in
the index.  If untracked is the only change, the stash fails and
the sync results in old stashes being applied instead.  The fix
causes git to also stash untracked changes and the resulting apply
works as expected.