Set gitflow branch repo configuration
+ On repo initialization, add include.path=../gitflow-branch-config
  to the repo's config file.
+ On self-update, initialize all repos to ensure that they have the
  proper gitflow branch configuration set.

Existing repos do not actually need to be updated more than once, nor is
self-update really an appriopriate place to perform this action.  Really
there should be a migration step which can be independently triggered.
However, given the rarity with which self-update is run, and the low
penalty for performing the repo update (~ 1.5 s/environment) this simple
solution was deemed preferable.