Update the installation guide
ZEN-33769.

The installation procedure for Docker was updated using the official
Docker installation guide for Ubuntu Linux. Now newdev-installer can be
executed on Ubuntu 16.04 LTS Xenial Xerus, as well as Ubuntu 18.04 LTS
Bionic Beaver without errors. The installation procedure for pip was
moved into the newdev-installer script. In addition, the procedure of
initialization of a zendev environment in the README file was updated by
adding some additional comments and in accordance with mentioned
changes.